+++ packages/flame/test/effects/effect_test_utils.dart
@@ -24,6 +24,7 @@ void effectTest(
   Vector2? expectedScale,
   required Random random,
 }) async {
+  component.children.register<ComponentEffect>();
   expectedPosition ??= Vector2.zero();
   expectedSize ??= Vector2.all(100.0);
   expectedScale ??= Vector2.all(1.0);