+++ packages/flame/CHANGELOG.md
@@ -54,6 +54,7 @@
  - `Game` is now a `Component`
  - `ComponentEffect` is now a `Component`
  - `HasGameRef` can now operate independently from `Game`
+ - `preOffset` and `postOffset` for effects to handle time before and after an effect
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps